chore(event): make `Event::from_json` private
